Simplifying the codes
