+++ src/pollution_model/pollution_tables.py
@@ -1,7 +1,10 @@
 """This script contains a SQLAlchemy model for the medusa 2.0 database table."""
 import abc
 
-from sqlalchemy import Column, Integer, String, Float
+from geoalchemy2 import Geometry
+from datetime import datetime, timezone
+
+from sqlalchemy import Column, Integer, String, Float, DateTime
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import DeclarativeMeta
 
@@ -156,3 +159,39 @@ class Medusa2ModelOutputRoads(_BaseMedusa2ModelOutput):
     __tablename__ = "medusa2_model_output_roads"
     geometry_table = "nz_roads"
     spatial_feature_id = Column(Integer, primary_key=True, name="road_id")
+
+
+class MedusaScenarios(Base):
+    """
+    Class representing the 'MedusaUserLogInfo' table.
+
+    Attributes
+    ----------
+    __tablename__ : str
+        Name of the database table.
+    scenario_id: int
+        Returns the model id of the new flood_model produce
+    antecedent_dry_days: float
+        The number of dry days between rainfall events.
+    average_rain_intensity: float
+        The intensity of the rainfall event in mm/h.
+    event_duration: float
+        The number of hours of the rainfall event.
+    rainfall_ph: float
+        The pH level of the rainfall, a measure of acidity.
+    created_at : datetime
+        Timestamp indicating when the log entry was created.
+    geometry : Polygon
+        Geometric representation of the catchment area coverage.
+    """  # pylint: disable=too-few-public-methods
+
+    __tablename__ = "medusa_scenarios"
+    scenario_id = Column(Integer, primary_key=True, autoincrement=True)
+
+    antecedent_dry_days = Column(Float, primary_key=True)
+    average_rain_intensity = Column(Float, primary_key=True)
+    event_duration = Column(Float, primary_key=True)
+    rainfall_ph = Column(Float, primary_key=True)
+
+    created_at = Column(DateTime(timezone=True), default=datetime.now(timezone.utc), comment="log created datetime")
+    geometry = Column(Geometry("POLYGON", srid=2193))
